services/websocket-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
confluent-kafka==2.6.1
redis==5.2.1
websockets==13.1
python-jose[cryptography]==3.3.0
